feat: Update keymap for Ferris keyboard
This commit updates the keymap for the Ferris keyboard, specifically for the mopemope keymap.  -
Added new control key definitions for equals (=) and minus (-) characters.  - Removed unused
keycodes CCR and CCW.  - Adjusted the COMBE and COMBO layers by replacing the removed keycodes with
XXXX (KC_NO).  - Replaced the unused placeholders in the home row combo layer with the new
equals (=) and minus (-) control keys.  - Removed the redundant definitions for CCR and CCW in the
`process_record_user` function.