chore(keymaps/totem/mopemope): replace TAB with GUI in QGMLWY layer
On the QGMLWY layer of the Totem Mopemope keymap, the left‐most key in the third row was changed
from TAB to GUI. This update provides an additional GUI modifier at the outer column and aligns the
layout with the intended command‐key placement.

an empty message aborts the commit.  # # On branch mopemope-totem # Your branch is up to date with
'mopemope/mopemope-totem'.  # # Changes to be committed: # modified:
keyboards/totem/keymaps/mopemope/keymap.c # # Untracked files: # .tool-versions # mopemope/ # totem/